# backend/internal/handler/v1/swagger/swagger.yaml
definitions:
...
  dd-prediction-api_internal_model.EditProjectReq:
    properties:
      name:
        type: string
      site_url:
        type: string
    type: object
...
  dd-prediction-api_internal_model.Project:
    properties:
      api_key:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_self_resolved_enabled:
        type: boolean
      is_users_duels_enabled:
        type: boolean
      name:
        type: string
      partner_id:
        type: string
      site_url:
        type: string
      status:
        $ref: '#/definitions/dd-prediction-api_internal_model.ProjectStatus'
      status_history:
        items:
          $ref: '#/definitions/dd-prediction-api_internal_model.ProjectStatusHistory'
        type: array
      updated_at:
        type: string
    type: object
...
  dd-prediction-api_internal_model.ProjectStatus:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - ProjectStatusActive
    - ProjectStatusDisabled
    - ProjectStatusBlocked
  dd-prediction-api_internal_model.ProjectStatusHistory:
    properties:
      changed_at:
        type: string
      id:
        type: integer
      project_id:
        type: string
      status:
        $ref: '#/definitions/dd-prediction-api_internal_model.ProjectStatus'
    type: object
...
  dd-prediction-api_internal_model.UpdateProjectPermissionsReq:
    properties:
      is_self_resolved_enabled:
        type: boolean
      is_users_duels_enabled:
        type: boolean
    type: object
...
  dd-prediction-api_pkg_apperrors.ErrorPublic:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a swagger specification for a Duel Duck back-end.
  title: Duel Duck Prediction API
  version: "1.0"
paths:
...
  /me/project:
    get:
      description: Returns the partner's project including name, site URL, status,
        and status history.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project details
          schema:
            $ref: '#/definitions/dd-prediction-api_internal_model.Project'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "403":
          description: Partner permission required
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Get own project
      tags:
      - project
    post:
      description: Creates a new project for the authenticated partner. Returns the
        project with its generated API key. One project per partner.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Created project
          schema:
            $ref: '#/definitions/dd-prediction-api_internal_model.Project'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "403":
          description: Partner permission required
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "409":
          description: Project already exists
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Create a partner project
      tags:
      - project
    put:
      consumes:
      - application/json
      description: Updates the project name and/or site URL.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Edit request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dd-prediction-api_internal_model.EditProjectReq'
      produces:
      - application/json
      responses:
        "200":
          description: Updated project
          schema:
            $ref: '#/definitions/dd-prediction-api_internal_model.Project'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "403":
          description: Partner permission required
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Edit partner project
      tags:
      - project
  /me/project/disable:
    put:
      description: Sets project status to disabled. API key will be rejected until
        the project is re-enabled. Cannot be used on admin-blocked projects.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project disabled
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "403":
          description: Forbidden — project is admin-blocked
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Disable partner project
      tags:
      - project
  /me/project/enable:
    put:
      description: Restores a partner-disabled project to active status. Does not
        affect admin-blocked projects.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project enabled
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "403":
          description: Forbidden — project is admin-blocked
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Re-enable partner project
      tags:
      - project
  /me/project/permissions:
    put:
      consumes:
      - application/json
      description: Enables or disables duel creation and self-resolve for users of
        this project's site. Self-resolve is automatically disabled when duel creation
        is disabled.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Permission flags
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dd-prediction-api_internal_model.UpdateProjectPermissionsReq'
      produces:
      - application/json
      responses:
        "200":
          description: Permissions updated
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "403":
          description: Partner permission required
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Update user permission flags
      tags:
      - project
...
securityDefinitions:
  BearerAuth:
    description: Enter the auth token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"